+++ webapp/src/components/inputs/_index.ts
@@ -1,0 +1,2 @@
+export { InputText } from "./text";
+export type { IInputTextProps } from "./text";
