+++ webapp/src/components/inputs/text.tsx
@@ -1,0 +1,5 @@
+export interface IInputTextProps {}
+
+export function InputText() {
+  return null
+}
