+++ webapp/src/components/meta/_index.ts
@@ -1,0 +1,8 @@
+export { blockComponent } from "./block-component";
+export { ClientComponent } from "./client-component";
+export type { IClientComponentProps } from "./client-component";
+export type {
+  IFuncComponent as FuncComponent,
+  IBlockProps,
+  IStylelessBlockProps,
+} from "./types";
